Subject: Revised EPUB packing flow to prioritize 'mimetype'
Changes in this commit mainly focus on the EPUB packing process within the EpubHandler class. The getFile and setFile functions are removed, and an enhancement in the pack function ensures 'mimetype' is the first, uncompressed file when packing the EPUB. Since some readers require the 'mimetype' file to be the first one in the EPUB archive, and uncompressed, this adjustment addresses that requirement, improving compatibility with a wider range of e-reader software.