Reordered build scripts and version bump in package.json
The command sequence in the build scripts within package.json has been altered to perform a version bump before building the application in major, minor, and patch builds. This counterintuitive order of steps was required to ensure the application's version number is updated before it is built. This change prevents situations where the application’s build does not reflect the correct, up-to-date version number.