Add success message functionality to Config.ts
A new string field 'successMessage' has been added to the Config class to store a generic success message. Similarly, an associated ELEMENT_ID has been defined. The 'updateAppContent' method now includes a call to a newly created method 'updateSuccessMessage'. This method sets the innerText of the success message element on the page to the value of 'successMessage'. The addition is to improve user feedback by displaying a success message upon successful file conversion.