Fix: Maintain 'readonly' modifier for version variable
The change corrects an irregularity that occurred while updating version in Config.ts. Previously, the 'readonly' modifier was being inadvertently omitted during version bump because the replacement string in regular expression didn't include it, hence modifying the 'version' property to be non-readonly.

This commit makes sure the 'readonly' modifier remains attached to the 'version' variable when updating the version number in the script, thus maintaining the immutability of the version property as intended. Thereby, upheld the consistency and robustness of the version control process.

No other consequential impacts stemming from this change.