Add second example to README
A new visual example has been added to the README to demonstrate how Bionic Reading enhances the reading experience, this time featuring the book "Algorithms to Live By". This further illustrates the capabilities of the application and provides the users with a broader understanding of its benefit. A relevant image for the example has also been added to the repository.